Shorten some test names
Paths to object files generated from them were too long and caused errors
