+++ src/test/run-pass/specialization/defaultimpl/allowed-cross-crate.rs
