+++ src/test/run-pass/specialization/defaultimpl/assoc-fns.rs
