+++ src/test/run-pass/specialization/defaultimpl/auxiliary/cross_crate_defaults.rs
