+++ src/test/run-pass/specialization/defaultimpl/basics.rs
