+++ src/test/run-pass/specialization/defaultimpl/default-methods.rs
