+++ src/test/run-pass/specialization/defaultimpl/overlap-projection.rs
