+++ src/test/run-pass/specialization/defaultimpl/projection-alias.rs
