+++ src/test/run-pass/specialization/defaultimpl/projection.rs
